Fix crash that prevents application from being closed
When a bookkeeping is closed, a new bookkeeping is created for the new
book year. The file path was not set before this commit, which lead to
a NullPointerException when Gogo Account was about to be exited. This
happened when the backup of the bookkeeping was created, which happens
when Gogo Account is exited.

Fixed by setting the file path when a new bookkeeping is created by
closing an existing bookkeeping.