Improve suggestion for account
It turns out that companies like Adyen take care of payments for many
companies. Payments all go to the same account of Adyen. When first
an account is suggested based on the account only, possibly a wrong
account can be suggested, because many companies now use Adyen's
account. So instead, first suggest an account based on the combination
of account and name, because the name of the company for whom the
payment is mentioned in the transaction.